requirements.txt
openai>=1.0.0
google-cloud-logging>=3.0.0
PyGithub>=2.0.0
python-dotenv>=1.0.0
